refactor: consistent type import
