+++ public/assets/images/icons/projectV2/ExploreIcon.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from '../../../../../src/features/common/types/common';
+import type { IconProps } from '../../../../../src/features/common/types/common';
 
 export const ExploreIcon = ({ width = '19px' }: IconProps) => {
   return (
