+++ public/assets/images/icons/projectV2/InfoIcon.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from '../../../../../src/features/common/types/common';
+import { type IconProps } from '../../../../../src/features/common/types/common';
 const InfoIcon = ({ width }: IconProps) => {
   return (
     <svg
