+++ public/assets/images/icons/projectV2/LocationIconOutline.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from '../../features/common/types/common';
+import { type IconProps } from '../../../../../src/features/common/types/common';
 
 const LocationIconOutline = ({ height, color }: IconProps) => {
   return (
