+++ public/assets/images/icons/projectV2/MailIcon.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IconProps } from '../../../../../src/features/common/types/common';
+import { type IconProps } from '../../../../../src/features/common/types/common';
 
 const MailIcon = ({ width, color }: IconProps) => {
   return (
