+++ public/assets/images/icons/projectV2/NewInfoIcon.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from '../../../../../src/features/common/types/common';
+import { type IconProps } from '../../../../../src/features/common/types/common';
 const NewInfoIcon = ({ width, color }: IconProps) => {
   return (
     <svg
