+++ public/assets/images/icons/projectV2/ViewProfileIcon.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IconProps } from '../../../../../src/features/common/types/common';
+import { type IconProps } from '../../../../../src/features/common/types/common';
 
 export const ViewProfileIcon = ({ width, color }: IconProps) => {
   return (
