+++ src/features/projectsV2/ProjectsMap/stories/ProjectSiteDropdown.stories.tsx
@@ -1,12 +1,12 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import ProjectSiteDropdown from '../ProjectSiteDropDown';
-import { Feature, Polygon } from 'geojson';
-import { SiteProperties } from '../ProjectSiteDropDown';
 import { useState } from 'react';
-import {
+import type { Meta, StoryObj } from '@storybook/react';
+import type { Feature, Polygon } from 'geojson';
+import type { SiteProperties } from '../ProjectSiteDropDown';
+import type {
   PlantLocation,
   SamplePlantLocation,
 } from '../../../common/types/plantLocation';
+import ProjectSiteDropdown from '../ProjectSiteDropDown';
 
 const meta: Meta<typeof ProjectSiteDropdown> = {
   title: 'Projects/Details/ProjectSiteDropdown',
@@ -127,170 +127,6 @@ const options: Feature<Polygon, SiteProperties>[] = [
     },
   },
 ];
-/* const selectedPlantLocation: PlantLocationMulti = {
-  nextMeasurementDate: null,
-  hid: 'ZAUYCK',
-  metadata: {
-    app: {
-      appVersion: '1.0.8',
-      deviceSystemName: 'Android',
-      deviceSystemVersion: '12',
-      deviceModel: 'SM-A536E',
-      deviceManufacturer: 'samsung',
-      deviceBrand: 'samsung',
-      deviceLocation: {
-        coordinates: [-90.1320438, 18.68243300000001],
-        type: 'Point',
-      },
-    },
-    public: [],
-  },
-  scientificName: null,
-  sampleInterventions: [
-    {
-      nextMeasurementDate: null,
-      parent: 'ivn_08FgSkuN0uLhQsTJkj1MFIJJ',
-      hid: 'EKHZIK',
-      metadata: {
-        app: {
-          appVersion: '1.0.8',
-          deviceSystemName: 'Android',
-          deviceSystemVersion: '12',
-          deviceModel: 'SM-A536E',
-          deviceManufacturer: 'samsung',
-          deviceBrand: 'samsung',
-          deviceLocation: {
-            coordinates: [-90.1318695, 18.6824257],
-            type: 'Point',
-          },
-        },
-        private: [],
-        public: [],
-      },
-      scientificName: 'Guazuma ulmifolia',
-      sampleInterventions: [],
-      description: null,
-      otherSpecies: null,
-      geometryUpdatesCount: 0,
-      type: 'sample-tree-registration',
-      interventionEndDate: '2022-08-17 00:00:00',
-      plantProjectSite: null,
-      statusReason: null,
-      registrationDate: '2022-08-17 00:00:00',
-      sampleTreeCount: null,
-      id: 'sivn_cMKoJoO5e9lm8sBlYysqES3O',
-      tag: '17810',
-      plantDate: '2022-08-17 00:00:00',
-      measurements: {
-        width: 0.1,
-        height: 0.4,
-      },
-      interventionStartDate: '2022-08-17 00:00:00',
-      idempotencyKey: '2b598c99b58fc9a41b3945c933353bf9',
-      profile: 'tpo_gEZeQNxNhxZZ54zvYzCofsCr',
-      coordinates: [
-        {
-          image: '6307f34b7a152996408413.jpg',
-          created: '2022-08-18 10:41:29',
-          coordinateIndex: 0,
-          id: 'coord_6vhswbkD5SFozbIyCCGPtPi7',
-          updated: '2022-08-25 22:10:19',
-          status: 'complete',
-        },
-      ],
-      scientificSpecies: 'sspec_sB8hEN129P1OTKPQMh',
-      history: [],
-      plantProject: 'xyz',
-      plantedSpecies: [],
-      originalGeometry: {
-        coordinates: [-90.1318695, 18.682425700000003],
-        type: 'Point',
-      },
-      captureMode: 'on-site',
-      geometry: {
-        coordinates: [-90.1318695, 18.682425700000003],
-        type: 'Point',
-      },
-      lastMeasurementDate: '2022-08-17 00:00:00',
-      captureStatus: 'complete',
-      deviceLocation: {
-        coordinates: [-90.1318695, 18.6824257],
-        type: 'Point',
-      },
-      status: null,
-    },
-  ],
-  description: null,
-  otherSpecies: null,
-  geometryUpdatesCount: 0,
-  type: 'multi-tree-registration',
-  interventionEndDate: '2022-08-17 00:00:00',
-  plantProjectSite: null,
-  statusReason: null,
-  registrationDate: '2022-08-17 00:00:00',
-  sampleTreeCount: 10,
-  id: 'ivn_08FgSkuN0uLhQsTJkj1MFIJJ',
-  tag: null,
-  plantDate: '2022-08-17 00:00:00',
-  measurements: null,
-  interventionStartDate: '2022-08-17 00:00:00',
-  idempotencyKey: 'b47f0c9c9f3dbb8210eaa921031f676e',
-  coordinates: [
-    {
-      image: '62fe159aecc95090182964.jpg',
-      coordinateIndex: 3,
-      id: 'coord_gqCtrzoo0fXz0KDn2urNOn29',
-      status: 'complete',
-    },
-  ],
-  scientificSpecies: null,
-  history: [],
-  plantProject: 'proj_WZkyugryh35sMmZMmXCwq7YY',
-  plantedSpecies: [
-    {
-      scientificName: 'Cordia dodecandra',
-      created: '2022-08-18 10:32:47',
-      otherSpecies: undefined,
-      scientificSpecies: 'sspec_0js3CX7f970bwh05mT',
-      treeCount: 486,
-      id: 'pspec_Ji94QImA2ArMRnYt5pukz4Ct',
-      updated: '2022-08-18 10:32:47',
-    },
-  ],
-  originalGeometry: {
-    coordinates: [
-      [
-        [-90.1320438, 18.68243300000001],
-        [-90.1313396, 18.6817704],
-        [-90.1309136, 18.682122000000003],
-        [-90.13161189999998, 18.68277789999999],
-        [-90.1320438, 18.68243300000001],
-      ],
-    ],
-    type: 'Polygon',
-  },
-  captureMode: 'on-site',
-  geometry: {
-    coordinates: [
-      [
-        [-90.1320438, 18.68243300000001],
-        [-90.1313396, 18.6817704],
-        [-90.1309136, 18.682122000000003],
-        [-90.13161189999998, 18.68277789999999],
-        [-90.1320438, 18.68243300000001],
-      ],
-    ],
-    type: 'Polygon',
-  },
-  lastMeasurementDate: null,
-  captureStatus: 'complete',
-  deviceLocation: {
-    coordinates: [-90.1320438, 18.68243300000001],
-    type: 'Point',
-  },
-  status: null,
-  image: null,
-}; */
 
 export const Preview: Story = {
   render: () => {
